+++ src/math/ray/ray.c
@@ -6,14 +6,14 @@
 /*   By: madelvin <madelvin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/20 18:01:54 by madelvin          #+#    #+#             */
-/*   Updated: 2025/03/23 16:18:56 by madelvin         ###   ########.fr       */
+/*   Updated: 2025/03/23 19:16:14 by madelvin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "structs.h"
 #include "maths.h"
 
-t_vec3	ray_at(t_ray r, double t)
+t_vec3	ray_at(t_ray r, float t)
 {
 	return ((t_vec3){
 		r.orig.x + t * r.dir.x,
