+++ src/math/vector/vec3_length.c
@@ -3,22 +3,22 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_length.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: madelvin <madelvin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/20 17:48:14 by madelvin          #+#    #+#             */
-/*   Updated: 2025/03/21 12:21:33 by lilefebv         ###   ########lyon.fr   */
+/*   Updated: 2025/03/23 19:16:14 by madelvin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "structs.h"
 #include <math.h>
 
-double	vec3_length_squared(const t_vec3 v)
+float	vec3_length_squared(const t_vec3 v)
 {
 	return (v.x * v.x + v.y * v.y + v.z * v.z);
 }
 
-double	vec3_length(const t_vec3 v)
+float	vec3_length(const t_vec3 v)
 {
 	return (sqrt(vec3_length_squared(v)));
 }
